Clarify instructions
As a beginner to the Python interpreter, I got a little confused.

I expected that the program will stop reading from sys.stdin
after I press Ctrl-Z, but it turned out that I had to press
the Enter key, after pressing Ctrl-Z.

Hopefully the change(s) in this commit will ease the user-experience
for someone else.